// packages/web/src/utils/enums.ts
export enum NOTIFICATION_TYPES {
  NEW_ARTICLE = 'NEW_ARTICLE',
  NEW_COMMENT = 'NEW_COMMENT',
  NEW_COMMENT_LIKE = 'NEW_COMMENT_LIKE',
  NEW_FOLLOWER = 'NEW_FOLLOWER',
  NEW_POST_LIKE = 'NEW_POST_LIKE',
  NEW_MENTION = 'NEW_MENTION',
  NEW_REPLY = 'NEW_REPLY',
  PRODUCT_ANNOUNCEMENTS = 'PRODUCT_ANNOUNCEMENTS',
  SIMILAR_PROJECTS = 'SIMILAR_PROJECTS',
}